新增被动合成仓，修复数字格式，修复超int卡死，修复eaep中键下单冲突

+++ src/main/java/com/sorrowmist/useless/content/recipe/ItemIngredientAllocator.java
@@ -107,7 +107,7 @@ public final class ItemIngredientAllocator {
         int low = 0;
         int high = upperLong > Integer.MAX_VALUE ? Integer.MAX_VALUE : (int) upperLong;
         while (low < high) {
-            int middle = low + (high - low + 1) / 2;
+            int middle = low + (int) (((long) high - low + 1L) / 2L);
             if (matches(requirements, inputs, middle)) {
                 low = middle;
             } else {
